+++ Flow/test/utopia/flow/test/collection/IteratorTest.scala
@@ -17,11 +17,11 @@ object IteratorTest extends App
 	assert(iter1.next() == 1)
 	assert(iter1.next() == 2)
 	assert(iter1.poll == 3)
-	assert(iter1.takeNextWhile { _ <= 10 }.size == 8)
+	assert(iter1.collectWhile { _ <= 10 }.size == 8)
 	assert(iter1.next() == 11)
 	assert(iter1.takeNext(10).size == 10)
 	assert(iter1.poll == 22)
-	iter1.takeNextWhile { _ < 100 }
+	iter1.collectWhile { _ < 100 }
 	assert(iter1.poll == 100)
 	iter1.next()
 	assert(!iter1.hasNext)